Solve the problem of instant chatting when switching chat rooms
