+++ app/channels/room_channel.rb
@@ -7,4 +7,5 @@ class RoomChannel < ApplicationCable::Channel
   def unsubscribed
     # Any cleanup needed when channel is unsubscribed
   end
+
 end
