+++ app/views/rooms/index.html.erb
@@ -1,4 +1,4 @@
-<div id="room-id" data-room-id="<%= @room.id%>"></div>
+<div id="room-id" data-room-id="<%= @room.id%>" data-controller="chatroom"></div>
 <div id="user-id" data-user-id="<%= current_user.id%>"></div>
   <div class="w-full h-screen px-20 overflow-y-scroll" id="chat-window">
       <% if @room.present? %>
